started to refactor the headers to remove a2denum.h and a2dobjs.h replaced with a2ddefs.h
