+++ README.md
@@ -15,9 +15,9 @@ A2D is a header only c++ templated library with python binding created with
     - [ ] decorate all functions that are needed from device
     - [ ] make all element operations in ```feelement.h``` support GPU parallelization
 - AD
-    - [ ] complete core operations
-    - [ ] complete AD/A2D expressions
-    - [ ] unit test for core operations and AD/A2D expressions
+    - [x] complete core operations
+    - [x] complete AD/A2D expressions
+    - [x] unit test for core operations and AD/A2D expressions
 - General
     - [ ] ```feelementmat.h``` add a parallel implementation of ElementMat (similar to ElementVector)
     - [ ] ```array.h``` replace reference implementations under ```BLAS``` namespace with Kokkos kernel functions
@@ -25,6 +25,9 @@ A2D is a header only c++ templated library with python binding created with
     - [ ] ```integrand_[poisson/elasticity/heat_conduction].h``` rename ```weak()``` -> ```residual()```
     - [ ] ```integrand_[poisson/elasticity/heat_conduction].h``` drop Jacobian vector product functor and Adjoint jacobian product functor, compute second order derivative matrices directly using ```jtransform()``` (see ```IntegrandTopoLinearElasticity::jacobian()``` for example)
     - [ ] ```feelement.h``` change ```add_jacobian()```, ```add_jacobian_vector_product()```, etc. accordingly
+- Matrix
+    - [ ] External library for parallel matrix solution on GPU
+    - [ ] External library for serial non-symmetric matrix factorization
